mb/clevo: Replace verb tables with reworked implementation
No modifications/assumptions about the codecs were made on these boards.

It should be noted that we do not modifiy the verb tables in any case,
as it would break the regression test script mentioned in the TEST
section below.

For an overall rationale for this rework, see commit 31fc5b06a6be
("device: Introduce reworked azalia verb table").

TEST= All boards passed regression test (CB:88763)

Change-Id: If505b6538293516a5813b47454c1407517894586
Signed-off-by: Nicholas Sudsgaard <devel+coreboot@nsudsgaard.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/88664
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Felix Singer <service+coreboot-gerrit@felixsinger.de>