mb/compulab: Replace verb tables with reworked implementation
Some boards did not provide the chip name for the audio codecs in the
comments, and were therefore identified using external sources:

  intense-pc:
    - 0x10ec0888 -> Realtek ALC888S-VC2[1]

It should be noted that we do not modifiy the verb tables in any case,
as it would break the regression test script mentioned in the TEST
section below.

For an overall rationale for this rework, see commit 31fc5b06a6be
("device: Introduce reworked azalia verb table").

References:
[1] Intense PC Hardware Specification, Revision 1.2 December 2013

TEST= All boards passed regression test (CB:88763)

Change-Id: I128a90fdba54937c148335753de3f65216b391ca
Signed-off-by: Nicholas Sudsgaard <devel+coreboot@nsudsgaard.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/88665
Reviewed-by: Felix Singer <service+coreboot-gerrit@felixsinger.de>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>