util/docker: Add jq for Jenkins nodes
Commit 48207895af70 ("lint: Warn about using change IDs for merged
changes") added a new lint test to warn about the usage of CB:XXXX for
already merged commits. However, 'jq' has to be installed in order to
run this test. As Jenkins currently does not have this dependency
installed, it prints the following message and skips the lint test
entirely:

> Check that CB:<change-id> does not refer to a merged change (lint-extended-025-merged-change-ids): This script requires "jq" to be installed

Fix this by installing the 'jq' dependency into Jenkins nodes.

Change-Id: Iae95ceabb09314c0c28239fdbd00f4e20df94254
Signed-off-by: Nicholas Sudsgaard <devel+coreboot@nsudsgaard.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/88920
Reviewed-by: Martin L Roth <gaumless@gmail.com>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Felix Singer <service+coreboot-gerrit@felixsinger.de>