amdblocks/lpc: Avoid decoding TPM MMIO to eSPI
Do not enable eSPI decoding for TPM MMIO resource. When the PC80
TPM PNP device was defined under the LPC/eSPI bridge scope, the
driver would detect it as a regular MMIO window to be opened on eSPI.
This caused the 0xFED40000-0xFED44FFF range to be enabled and routed
to eSPI and broke the SPI TPM communication in ramstage. Explicitly
enable SPI TPM decoding with lpc_tpm_decode_spi in such case to avoid
the incorrect eSPI MMIO decode programming.

Add console prints to debug which resources are actually decoded, and
to detect such board configuration issues earlier in the future.

TEST=SPI TPM startup does not fail on Gigabyte MZ33-AR1.

Change-Id: If9763218896f3587a2b76c7eb36b811fe5da7a10
Signed-off-by: Michał Żygowski <michal.zygowski@3mdeb.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/89193
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Felix Held <felix-coreboot@felixheld.de>