mb/purism: add missing terminators to azalia codec tables
This was supposed to be checked in the regression test script
(CB:88763), however it turns out Valgrind's Memcheck only works on the
heap memory and is unable to catch such errors.

The regression test script was modified to use AddressSanitizer which
can catch such errors, so this should not be a problem in subsequent
changes during the verb table rework.

To be safe, the previously merged commits were also checked with the new
regression test script:

  f634121fa476 ("mb/purism: Replace verb tables with reworked implementation")
  20d404245832 ("mb/asrock: Replace verb tables with reworked implementation")
  2b7dbf80c981 ("mb/apple: Replace verb tables with reworked implementation")
  970249694f85 ("mb/amd: Replace verb tables with reworked implementation")
  94beaa7ab313 ("mb/acer: Replace verb tables with reworked implementation")
  f3db3a19d5cb ("mb/51nb: Replace verb tables with reworked implementation")

However, the following mini-HD code was checked manually, as figuring
out how to strip out minihd_init() was not worth the effort:

  bc92d9a66666 ("nb/intel/haswell/minihd.c: Add reworked verb table implementation")
  69781b980680 ("soc/intel/broadwell/minihd.c: Add reworked verb table implementation")

Change-Id: Iea964fb8b92814b57d4c82412c47cf31fa48de66
Signed-off-by: Nicholas Sudsgaard <devel+coreboot@nsudsgaard.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/89376
Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>