soc/amd/turin_poc: Add hook for IOAPIC interrupt routing
AMD systems have a bit in IOAPIC private configuration space that
allow redirecting the masked interrupts to the FCH IOAPIC/PIC. When
the OS decides to use IOAPIC mode for interrupt routing, this bit
has to be cleared on all IOAPICs.

Implement Turin-specific code that clears the aforementioned bit
when OS decides to use IOAPIC for interrupt routing.

Change-Id: Ibbd436049b7850c4640195bebe5a48873a5f9664
Signed-off-by: Michał Żygowski <michal.zygowski@3mdeb.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/89480
Reviewed-by: Alicja Michalska <ahplka19@gmail.com>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>